labelling_pipeline/yolo_api/requirements.txt
fastapi
uvicorn
boto3
ultralytics
python-dotenv
Pillow
opencv-python
craft-text-detector